Script to get dictionary with the mappings
Not the best, but it should be good for now. Need to make it more
robust. There are also edge cases that I just defaulted to the original
token without '_'. I assumed the API rate limit was 3 as that was one of
the lowest ones I've seen. Could make semaphores work differently based
on different APIs but decided not to.